+++ examples/generative/corrdiff/conf/dataset/cwb_generate.yaml
@@ -1,3 +1,4 @@
+
 # SPDX-FileCopyrightText: Copyright (c) 2023 - 2024 NVIDIA CORPORATION & AFFILIATES.
 # SPDX-FileCopyrightText: All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
@@ -14,28 +15,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-hydra:
-  job:
-    chdir: true
-    name: regression
-  run:
-    dir: ./outputs/${hydra:job.name}
-
-# Get defaults
-defaults:
-
-  # Dataset
-  - dataset/cwb_train
-
-  # Model
-  - model/corrdiff_regression
-  #- model/corrdiff_diffusion
-  #- model/corrdiff_patched_diffusion
-
-  # Training
-  - training/corrdiff_regression
-  #- training/corrdiff_diffusion
-  #- training/corrdiff_patched_diffusion
-
-  # Validation
-  - validation/basic
+type: cwb
+train: False
+all_times: True
+n_history: 0
+in_channels: [0, 1, 2, 3, 4, 9, 10, 11, 12, 17, 18, 19]
+out_channels: [0, 17, 18, 19]
+img_shape_x: 448
+img_shape_y: 448
+roll: false
+add_grid: true
+ds_factor: 1
+min_path: null
+max_path: null
+global_means_path: null
+global_stds_path: null
+normalization: "v1"
