+++ examples/generative/corrdiff/conf/generation/base.yaml
@@ -1,0 +1,79 @@
+# SPDX-FileCopyrightText: Copyright (c) 2023 - 2024 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-FileCopyrightText: All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+seed_batch_size: 32
+  # Size of the batched inference
+inference_mode: regression_and_diffusion
+patch_size: 448
+patch_shape_x: 448
+patch_shape_y: 448
+  # Patch size. Patch-based sampling will be utilized if these dimensions differ from 
+  # img_shape_x and img_shape_y
+overlap_pixels: 0
+  # Number of overlapping pixels between adjacent patches
+boundary_pixels: 0
+  # Number of boundary pixels to be cropped out. 2 is recommanded to address the boundary
+  # artifact.
+hr_mean_conditioning: false
+gridtype: "sinusoidal"
+N_grid_channels: 4
+sample_res: full
+  # Sampling resolution
+times_range: null
+times:
+  - 2021-02-02T00:00:00
+  - 2021-03-02T00:00:00
+  - 2021-04-02T00:00:00
+  # hurricane
+  - 2021-09-12T00:00:00
+  - 2021-09-12T12:00:00
+
+perf:
+  force_fp16: False
+    # Whether to force fp16 precision for the model. If false, it'll use the precision
+    # specified upon training.
+  use_torch_compile: false
+    # whether to use torch.compile on the diffusion model
+    # this will make the first time stamp generation very slow due to compilation overheads
+    # but will significantly speed up subsequent inference runs
+  num_writer_workers: 1
+    # number of workers to use for writing file
+    # To support multiple workers a threadsafe version of the netCDF library must be used
+
+io:
+  res_ckpt_filename: diffusion_checkpoint.mdlus
+    # Checkpoint filename for the diffusion model  
+  reg_ckpt_fimename: regression_checkpoint.mdlus
+    # Checkpoint filename for the mean predictor model
+
+dataset:
+  type: cwb
+  train: False
+  all_times: True
+  data_path: /data/data2/2023-01-24-cwb-4years.zarr
+  n_history: 0
+  in_channels: [0, 1, 2, 3, 4, 9, 10, 11, 12, 17, 18, 19]
+  out_channels: [0, 17, 18, 19]
+  img_shape_x: 448
+  img_shape_y: 448
+  roll: false
+  add_grid: true
+  ds_factor: 4
+  min_path: null
+  max_path: null
+  global_means_path: null
+  global_stds_path: null
+  normalization: "v1"
