+++ examples/generative/corrdiff/conf/sampler/deterministic.yaml
@@ -14,28 +14,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-hydra:
-  job:
-    chdir: true
-    name: regression
-  run:
-    dir: ./outputs/${hydra:job.name}
 
-# Get defaults
-defaults:
-
-  # Dataset
-  - dataset/cwb_train
-
-  # Model
-  - model/corrdiff_regression
-  #- model/corrdiff_diffusion
-  #- model/corrdiff_patched_diffusion
-
-  # Training
-  - training/corrdiff_regression
-  #- training/corrdiff_diffusion
-  #- training/corrdiff_patched_diffusion
-
-  # Validation
-  - validation/basic
+type: deterministic
+num_steps: 9
+  # Number of denoising steps
+num_ensembles: 64
+  # Number of ensembles to generate per input
+solver: euler
+  # ODE solver [euler, heun]
